removes words that may be inappropriate; also removes duplicates (#1)
For the nouns list there are a few duplicates and most of them are inappropriate words, more or less. Some of them may seem innocent (e.g. "ball" or "bottom") but they can lead to inappropriate meanings in combination with a lot of adjectives.

For the adjectives list it is the other way around. There are a few ones that I took out like "dirty" of "filthy" because they make everything sound dirty or filthy :)
But the majority of them (~98-99%) are just duplicates.